HZN-1202: Moved collection base classes into collection/core project
Removed dependencies on opennms-services from features/collection
projects. Instead, opennms-services will depend on these projects and
will only contain the SNMP-specific collection classes and the
Collectd daemon/scheduler.
