+++ features/collection/api/src/main/java/org/opennms/netmgt/collection/api/CollectionAgentService.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.collectd;
+package org.opennms.netmgt.collection.api;
 
 import java.net.InetAddress;
