+++ features/collection/api/src/main/java/org/opennms/netmgt/collection/api/CollectionTimedOut.java
@@ -1,8 +1,8 @@
 /*******************************************************************************
  * This file is part of OpenNMS(R).
  *
- * Copyright (C) 2017-2017 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2017 The OpenNMS Group, Inc.
+ * Copyright (C) 2002-2014 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2014 The OpenNMS Group, Inc.
  *
  * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
  *
@@ -26,26 +26,24 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.collectd;
-
-import org.opennms.netmgt.collection.api.CollectionException;
+package org.opennms.netmgt.collection.api;
 
 /**
- * This exception should be thrown by a {@link ServiceCollector} when
- * collection fails, but should not trigger a status change (resulting
- * in an alarm) for target service.
+ * <p>CollectionTimedOut class.</p>
  *
- * For example, collection may fail if the OpenNMS system is in the process
- * of shutting down, and required components are not available.
- *
- * @author jesse
+ * @author ranger
+ * @version $Id: $
  */
-public class CollectionUnknown extends CollectionException {
+public class CollectionTimedOut extends CollectionException {
 
-    private static final long serialVersionUID = 5630156328994222706L;
+    private static final long serialVersionUID = 7277882701768744179L;
 
-    public CollectionUnknown(String message, Throwable cause) {
-        super(message, cause);
+    /**
+     * <p>Constructor for CollectionTimedOut.</p>
+     *
+     * @param message a {@link java.lang.String} object.
+     */
+    public CollectionTimedOut(String message) {
+        super(message);
     }
-
 }
