+++ features/collection/core/src/main/java/org/opennms/netmgt/collection/core/DefaultCollectionAgentFactory.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.collectd;
+package org.opennms.netmgt.collection.core;
 
 import java.net.InetAddress;
