+++ opennms-services/src/main/java/org/opennms/netmgt/collectd/CollectableService.java
@@ -29,8 +29,8 @@
 package org.opennms.netmgt.collectd;
 
 import java.io.File;
-import java.util.Date;
 import java.net.InetAddress;
+import java.util.Date;
 
 import org.opennms.core.logging.Logging;
 import org.opennms.core.utils.InetAddressUtils;
@@ -39,14 +39,20 @@ import org.opennms.netmgt.collection.api.AttributeGroup;
 import org.opennms.netmgt.collection.api.CollectionAgent;
 import org.opennms.netmgt.collection.api.CollectionAttribute;
 import org.opennms.netmgt.collection.api.CollectionException;
+import org.opennms.netmgt.collection.api.CollectionFailed;
 import org.opennms.netmgt.collection.api.CollectionInitializationException;
 import org.opennms.netmgt.collection.api.CollectionResource;
 import org.opennms.netmgt.collection.api.CollectionSet;
 import org.opennms.netmgt.collection.api.CollectionSetVisitor;
 import org.opennms.netmgt.collection.api.CollectionStatus;
+import org.opennms.netmgt.collection.api.CollectionTimedOut;
+import org.opennms.netmgt.collection.api.CollectionUnknown;
+import org.opennms.netmgt.collection.api.CollectionWarning;
 import org.opennms.netmgt.collection.api.PersisterFactory;
 import org.opennms.netmgt.collection.api.ServiceParameters;
 import org.opennms.netmgt.collection.api.TimeKeeper;
+import org.opennms.netmgt.collection.core.CollectionSpecification;
+import org.opennms.netmgt.collection.core.DefaultCollectionAgent;
 import org.opennms.netmgt.collection.support.AttributeGroupWrapper;
 import org.opennms.netmgt.collection.support.CollectionAttributeWrapper;
 import org.opennms.netmgt.collection.support.CollectionResourceWrapper;
@@ -190,7 +196,7 @@ class CollectableService implements ReadyRunnable {
     /**
      * <p>getSpecification</p>
      *
-     * @return a {@link org.opennms.netmgt.collectd.CollectionSpecification} object.
+     * @return a {@link org.opennms.netmgt.collection.core.CollectionSpecification} object.
      */
     public CollectionSpecification getSpecification() {
     	return m_spec;
