+++ opennms-services/src/main/resources/META-INF/opennms/applicationContext-collectd.xml
@@ -23,7 +23,7 @@
       <property name="eventIpcManager" ref="eventIpcManager"/>
     </bean>
 
-    <bean id="collectionAgentFactory" class="org.opennms.netmgt.collectd.DefaultCollectionAgentFactory" />
+    <bean id="collectionAgentFactory" class="org.opennms.netmgt.collection.core.DefaultCollectionAgentFactory" />
     <onmsgi:service interface="org.opennms.netmgt.collection.api.CollectionAgentFactory" ref="collectionAgentFactory" />
 
     <bean id="defaultResourceTypeMapper" class="org.opennms.netmgt.collectd.DefaultResourceTypeMapper" />
