+++ opennms-services/src/test/java/org/opennms/netmgt/collectd/CollectdIntegrationTest.java
@@ -50,6 +50,7 @@ import org.junit.Test;
 import org.junit.rules.TestName;
 import org.opennms.core.utils.InsufficientInformationException;
 import org.opennms.netmgt.collection.support.DefaultServiceCollectorRegistry;
+import org.opennms.netmgt.collection.test.api.CollectorTestUtils;
 import org.opennms.netmgt.config.CollectdConfigFactory;
 import org.opennms.netmgt.config.PollOutagesConfigFactory;
 import org.opennms.netmgt.config.ThresholdingConfigFactory;
