+++ opennms-services/src/test/resources/org/opennms/netmgt/config/http-datacollection-config-NMS4886.xml
