+++ opennms-services/src/test/resources/org/opennms/netmgt/config/http-datacollection-config.xml
