+++ pom.xml
@@ -1998,6 +1998,11 @@
       </dependency>
       <dependency>
         <groupId>org.opennms.features.collection</groupId>
+        <artifactId>org.opennms.features.collection.core</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.opennms.features.collection</groupId>
         <artifactId>org.opennms.features.collection.client-rpc</artifactId>
         <version>${project.version}</version>
       </dependency>
